Added something

+++ JGamePackage/JGame/Classes/UI/Modifiers/UIBorder.java
@@ -3,7 +3,7 @@ package JGamePackage.JGame.Classes.UI.Modifiers;
 import java.awt.Color;
 
 public class UIBorder extends UIModifier {
-    public double Width = 5;
+    public int Width = 5;
 
     public Color BorderColor = Color.black;
